+++ packages/sdk-metrics/src/state/MetricStorage.ts
@@ -55,6 +55,7 @@ export abstract class MetricStorage {
         description: description,
         valueType: this._instrumentDescriptor.valueType,
         unit: this._instrumentDescriptor.unit,
+        advice: this._instrumentDescriptor.advice,
       }
     );
   }